[태그] 태그 삭제/수정 기능 추가, AnimationState 옮김, dp -> px 함수 중복 제거

+++ commonBase/src/main/java/com/lok/dev/commonbase/BaseDialogFragment.kt
@@ -119,6 +119,6 @@ abstract class BaseDialogFragment<Binding : ViewDataBinding, Result> : DialogFra
         this.setOnClickListener { onDialogBackPressed() }
     }
 
-    val Number.dp: Float get() = requireContext().convertDpToPx(this.toFloat())
+    val Number.dp: Int get() = requireContext().convertDpToPx(this.toFloat()).toInt()
 
 }
